+++ zcash_client_backend/CHANGELOG.md
@@ -18,6 +18,8 @@ and this library adheres to Rust's notion of
     - `LightdInfo::estimated_height`
     - `LightdInfo::donation_address`
 - `zcash_client_backend::tor`:
+  - `Client::http_get`
+  - `Client::http_get_json`
   - `Client::set_dormant`
   - `DormantMode`